requirements.txt
apache-beam==2.27.0
apache-flink==1.13.1
apache-flink-libraries==1.13.1
avro-python3==1.9.2.1
certifi==2021.5.30
chardet==4.0.0
cloudpickle==1.2.2
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
fastavro==0.23.6
future==0.18.2
grpcio==1.38.0
hdfs==2.6.0
httplib2==0.17.4
idna==2.10
kafka-python==2.0.2
mock==2.0.0
numpy==1.19.5
oauth2client==4.1.3
pandas==1.1.5
pbr==5.6.0
protobuf==3.17.2
py4j==0.10.8.1
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.0
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
typing-extensions==3.7.4.3
urllib3==1.26.5
